+++ Plugins/Rhetos.LightDMS/Download/DownloadHelper.cs
@@ -17,9 +17,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Features;
-using Microsoft.AspNetCore.StaticFiles;
+using Amazon.S3.Model;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Blob;
 using Rhetos.LightDms.Storage;
@@ -28,8 +26,13 @@ using Rhetos.Utilities;
 using System;
 using System.Data;
 using System.Data.SqlClient;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Net;
+using System.Net.Security;
+using System.Security.Cryptography.X509Certificates;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
 
@@ -40,112 +43,153 @@ namespace Rhetos.LightDMS
         private const int BUFFER_SIZE = 100 * 1024; // 100 kB buffer
 
         private readonly ILogger _logger;
-        private readonly ConnectionString _connectionString;
-        private readonly IContentTypeProvider _contentTypeProvider;
-        private readonly Respond _respond;
-        private readonly LightDMSOptions _lightDMSOptions;
+        private readonly bool _detectResponseBlockingErrors;
+        private readonly int _detectResponseBlockingErrorsTimeoutMs;
 
-        public DownloadHelper(ILogProvider logProvider,
-            ConnectionString connectionString,
-            IContentTypeProvider contentTypeProvider,
-            LightDMSOptions lightDMSOptions)
+        public DownloadHelper()
         {
-            _connectionString = connectionString;
+            var logProvider = new NLogProvider();
             _logger = logProvider.GetLogger(GetType().Name);
-            _contentTypeProvider = contentTypeProvider;
-            _lightDMSOptions = lightDMSOptions;
-            _respond = new Respond(logProvider);
+            var configuration = new Utilities.Configuration();
+            _detectResponseBlockingErrors = configuration.GetBool("LightDMS.DetectResponseBlockingErrors", true).Value;
+            _detectResponseBlockingErrorsTimeoutMs = configuration.GetInt("LightDMS.DetectResponseBlockingErrorsTimeoutMs", 60 * 1000).Value;
         }
 
-        public async Task HandleDownload(HttpContext context, Guid? documentVersionId, Guid? fileContentId)
+        public void HandleDownload(HttpContext context, Guid? documentVersionId, Guid? fileContentId)
         {
             try
             {
-                using (var sqlConnection = new SqlConnection(_connectionString))
+                using (var sqlConnection = new SqlConnection(SqlUtility.ConnectionString))
                 {
                     sqlConnection.Open();
-                    var fileMetadata = GetMetadata(context, documentVersionId, fileContentId, sqlConnection);
+                    var fileMetadata = GetFileMetadata(documentVersionId, fileContentId, sqlConnection, GetFileNameFromQueryString(context));
 
-                    context.Response.StatusCode = (int)HttpStatusCode.OK;
+                    PopulateHeader(context, fileMetadata.FileName, fileMetadata.Size);
 
-                    if (!await TryDownloadFromAzureBlob(context, fileMetadata))
-                    {
-                        //If there is no document on AzureBlobStorage, take it from DB
-                        context.Features.Get<IHttpResponseBodyFeature>()?.DisableBuffering();
-                        if (!await TryDownloadFromFileStream(context, fileMetadata, sqlConnection))
-                            // If FileStream is not available - read from VarBinary(MAX) column using buffer;
-                            await DownloadFromVarbinary(context, fileMetadata, sqlConnection);
-                    }
+                    ResolveDownload(fileMetadata, sqlConnection, context.Response.OutputStream, context.Response, context);
                 }
             }
             catch (Exception ex)
             {
                 if (ex.Message == "Function PathName is only valid on columns with the FILESTREAM attribute.")
-                    await _respond.BadRequest(context, "FILESTREAM attribute is missing from LightDMS.FileContent.Content column. However, file is still available from download via REST interface.");
+                    Respond.BadRequest(context, "FILESTREAM attribute is missing from LightDMS.FileContent.Content column. However, file is still available from download via REST interface.");
                 else
-                    await _respond.InternalError(context, ex);
+                    Respond.InternalError(context, ex);
             }
         }
 
-        class FileMetadata
+        private void Download(FileMetadata fileDownloadMetadata, HttpContext context, Stream stream)
         {
-            public Guid FileContentId;
-            public string FileName;
-            public bool AzureStorage;
-            public long Size;
+            context.Response.AddHeader("Content-Length", fileDownloadMetadata.Size.ToString());
+
+            var buffer = new byte[BUFFER_SIZE];
+            int bytesRead;
+
+            int totalBytesWritten = 0;
+            while ((bytesRead = stream.Read(buffer, 0, BUFFER_SIZE)) > 0)
+            {
+                if (!context.Response.IsClientConnected)
+                    break;
+
+                void writeResponse() => context.Response.OutputStream.Write(buffer, 0, bytesRead);
+
+                if (_detectResponseBlockingErrors)
+                {
+                    // HACK: `Response.OutputStream.Write` sometimes blocks the process at System.Web.dll!System.Web.Hosting.IIS7WorkerRequest.ExplicitFlush();
+                    // Until the issue is solved, this hack allows 1. logging of the problem, and 2. closing the SQL transaction while the thread remains blocked.
+                    // Tried removing PreSendRequestHeaders, setting aspnet:UseTaskFriendlySynchronizationContext and disabling anitivirus, but it did not help.
+                    // Total performance overhead of Task.Run...Wait is around 0.01 sec when downloading 200MB file with BUFFER_SIZE 100kB.
+                    var task = Task.Run(writeResponse);
+                    if (!task.Wait(_detectResponseBlockingErrorsTimeoutMs))
+                    {
+                        throw new FrameworkException(ResponseBlockedMessage +
+                            $" Process {Process.GetCurrentProcess().Id}, thread {Thread.CurrentThread.ManagedThreadId}," +
+                            $" streamed {totalBytesWritten} bytes of {fileDownloadMetadata.Size}, current batch {bytesRead} bytes.");
+                    }
+                }
+                else
+                    writeResponse();
+
+                totalBytesWritten += bytesRead;
+                context.Response.Flush();
+            }
         }
 
-        private FileMetadata GetMetadata(HttpContext context, Guid? documentVersionId, Guid? fileContentId, SqlConnection sqlConnection)
+        public void ResolveDownload(FileMetadata fileMetadata, SqlConnection sqlConnection, Stream outputStream = null, HttpResponse httpResponse = null, HttpContext context = null)
         {
-            // if "filename" is present in query, that one is used as download filename
-            var query = context.Request.Query;
-            string queryFileName = null;
-            foreach (var key in query.Keys) if (key.ToLower() == "filename") queryFileName = query[key];
+            if (fileMetadata.S3Storage)
+                DownloadFromS3(fileMetadata, context);
+            else if (fileMetadata.AzureStorage)
+                DownloadFromAzureBlob(fileMetadata.FileContentId, outputStream, httpResponse);
+            else if (IsFileStream(sqlConnection))
+                DownloadFromFileStream(fileMetadata, sqlConnection, context);
+            else
+                DownloadFromVarbinary(fileMetadata, sqlConnection, context);
+        }
 
-            SqlCommand getMetadata = null;
+        private bool IsFileStream(SqlConnection sqlConnection)
+        {
+            using (var sqlCommand = new SqlCommand("SELECT TOP 1 1 FROM sys.columns c WHERE OBJECT_SCHEMA_NAME(C.object_id) = 'LightDMS' AND OBJECT_NAME(C.object_id) = 'FileContent' AND c.Name = 'Content' AND c.is_filestream = 1", sqlConnection))
+            {
+                return sqlCommand.ExecuteScalar() != null;
+            }
+        }
+
+        private static string GetFileNameFromQueryString(HttpContext context)
+        {
+            var query = HttpUtility.ParseQueryString(context.Request.Url.Query);
+            string queryFileName = null;
+            foreach (var key in query.AllKeys) if (key.ToLower() == "filename") queryFileName = query[key];
+
+            return queryFileName;
+        }
+
+        private FileMetadata GetFileMetadata(Guid? documentVersionId, Guid? fileContentId, SqlConnection sqlConnection, string queryStringFileName)
+        {
+            SqlCommand getFileMetadata;
             if (documentVersionId != null)
-                getMetadata = new SqlCommand(@"
+                getFileMetadata = new SqlCommand(@"
                         SELECT
                             dv.FileName,
                             FileSize = DATALENGTH(Content),
                             dv.FileContentID,
-                            fc.AzureStorage
+                            fc.AzureStorage,
+                            fc.S3Storage
                         FROM
                             LightDMS.DocumentVersion dv
                             INNER JOIN LightDMS.FileContent fc ON dv.FileContentID = fc.ID
                         WHERE 
                             dv.ID = '" + documentVersionId + @"'", sqlConnection);
             else
-                getMetadata = new SqlCommand(@"
+                getFileMetadata = new SqlCommand(@"
                         SELECT 
                             FileName ='unknown.txt',
                             FileSize = DATALENGTH(Content),
                             FileContentID = fc.ID,
-                            AzureStorage = CAST(0 AS BIT)
+                            AzureStorage = CAST(0 AS BIT),
+                            S3Storage = CAST(0 AS BIT)
                         FROM 
                             LightDMS.FileContent fc 
                         WHERE 
                             ID = '" + fileContentId + "'", sqlConnection);
 
-            using (var result = getMetadata.ExecuteReader(CommandBehavior.SingleRow))
+            using (var result = getFileMetadata.ExecuteReader(CommandBehavior.SingleRow))
             {
                 result.Read();
                 return new FileMetadata
                 {
                     FileContentId = (Guid)result["FileContentID"],
-                    FileName = queryFileName ?? (string)result["FileName"],
-                    AzureStorage = result["AzureStorage"] != DBNull.Value ? (bool)result["AzureStorage"] : false,
-                    Size = (long)result["FileSize"],
+                    FileName = queryStringFileName ?? (string)result["FileName"],
+                    AzureStorage = result["AzureStorage"] != DBNull.Value && (bool)result["AzureStorage"],
+                    S3Storage = result["S3Storage"] != DBNull.Value && (bool)result["S3Storage"],
+                    Size = (long)result["FileSize"]
                 };
             }
         }
 
-        private async Task<bool> TryDownloadFromAzureBlob(HttpContext context, FileMetadata file)
+        private void DownloadFromAzureBlob(Guid fileContentId, Stream outputStream, HttpResponse httpResponse)
         {
-            if (file.AzureStorage == false)
-                return false;
-
-            var storageConnectionVariable = _lightDMSOptions.StorageConnectionVariable;
+            var storageConnectionVariable = System.Configuration.ConfigurationManager.AppSettings.Get("LightDms.StorageConnectionVariable");
             string storageConnectionString;
             if (!string.IsNullOrWhiteSpace(storageConnectionVariable))
                 storageConnectionString = Environment.GetEnvironmentVariable(storageConnectionVariable, EnvironmentVariableTarget.Machine);
@@ -155,146 +199,122 @@ namespace Rhetos.LightDMS
 
             if (!string.IsNullOrEmpty(storageConnectionString))
             {
-                CloudStorageAccount storageAccount;
-                if (!CloudStorageAccount.TryParse(storageConnectionString, out storageAccount))
+                if (!CloudStorageAccount.TryParse(storageConnectionString, out CloudStorageAccount storageAccount))
                     throw new FrameworkException("Invalid Azure Blob Storage connection string.");
 
                 CloudBlobClient cloudBlobClient = storageAccount.CreateCloudBlobClient();
-                var storageContainerName = _lightDMSOptions.StorageContainer;
+                var storageContainerName = System.Configuration.ConfigurationManager.AppSettings.Get("LightDms.StorageContainer");
                 if (string.IsNullOrWhiteSpace(storageContainerName))
                     throw new FrameworkException("Azure blob storage container name is missing from configuration.");
 
                 CloudBlobContainer cloudBlobContainer = cloudBlobClient.GetContainerReference(storageContainerName);
-                if (!await cloudBlobContainer.ExistsAsync())
+                if (!cloudBlobContainer.Exists())
                     throw new FrameworkException("Azure blob storage container doesn't exist.");
 
-                CloudBlockBlob cloudBlockBlob = cloudBlobContainer.GetBlockBlobReference("doc-" + file.FileContentId.ToString());
-                if (await cloudBlockBlob.ExistsAsync())
+                CloudBlockBlob cloudBlockBlob = cloudBlobContainer.GetBlockBlobReference("doc-" + fileContentId.ToString());
+                if (cloudBlockBlob.Exists())
                 {
                     try
                     {
-                        await cloudBlockBlob.FetchAttributesAsync();
+                        cloudBlockBlob.FetchAttributes();
 
-                        PopulateHeader(context, file.FileName, cloudBlockBlob.Properties.Length);
-                        await cloudBlockBlob.DownloadToStreamAsync(context.Response.Body);
+                        if (httpResponse != null)
+                            httpResponse.AddHeader("Content-Length", cloudBlockBlob.Properties.Length.ToString());
 
-                        await context.Response.Body.FlushAsync();
-                        return true;
+                        // Downloads directly to outputStream
+                        cloudBlockBlob.DownloadToStream(outputStream);
                     }
                     catch (Exception ex)
                     {
-                        //when unexpected error occurs log it, then fall back to DB
-                        _logger.Error("Azure storage error, falling back to DB. Error: " + ex.ToString());
-                        return false;
+                        _logger.Error("Azure storage error. Error: " + ex.ToString());
                     }
                 }
-                else
-                    return false; //if no blob is present fall back to DB
             }
             else
                 throw new FrameworkException("Azure Blob Storage environment variable missing.");
         }
 
-        private async Task<bool> TryDownloadFromFileStream(HttpContext context, FileMetadata file, SqlConnection sqlConnection)
+        private void DownloadFromS3(FileMetadata fileMetadata, HttpContext context)
         {
-            SqlCommand checkFileStreamEnabled = new SqlCommand("SELECT TOP 1 1 FROM sys.columns c WHERE OBJECT_SCHEMA_NAME(C.object_id) = 'LightDMS' AND OBJECT_NAME(C.object_id) = 'FileContent' AND c.Name = 'Content' AND c.is_filestream = 1", sqlConnection);
-            if (checkFileStreamEnabled.ExecuteScalar() == null)
-                return false;
-
-            SqlTransaction sqlTransaction = sqlConnection.BeginTransaction(IsolationLevel.ReadCommitted); // Explicit transaction is required when working with SqlFileStream class.
-            try
-            {
-                using (var fileStream = SqlFileStreamProvider.GetSqlFileStreamForDownload(file.FileContentId, sqlTransaction))
-                {
-                    byte[] buffer = new byte[BUFFER_SIZE];
-                    int bytesRead;
-                    PopulateHeader(context, file.FileName, file.Size);
-                    int totalBytesWritten = 0;
-                    while ((bytesRead = fileStream.Read(buffer, 0, BUFFER_SIZE)) > 0)
+            ServicePointManager.ServerCertificateValidationCallback +=
+                    delegate (
+                        object sender,
+                        X509Certificate certificate,
+                        X509Chain chain,
+                        SslPolicyErrors sslPolicyErrors)
                     {
-                        if (context.RequestAborted.IsCancellationRequested)
-                            break;
+                        if (certificate.Subject.IndexOf("ssc.gov.hr") > -1)
+                            return true;
+                        return sslPolicyErrors == SslPolicyErrors.None;
+                    };
+            
+            using (var client = S3StorageClient.GetAmazonS3Client())
+            {
+                GetObjectRequest getObjRequest = new GetObjectRequest();
+                getObjRequest.BucketName = ConfigUtility.GetAppSetting("LightDms.S3.BucketName");
+                if (string.IsNullOrWhiteSpace(getObjRequest.BucketName))
+                    throw new FrameworkException("Missing S3 storage bucket name.");
+                
+                var s3Folder = ConfigUtility.GetAppSetting("LightDms.S3.DestinationFolder");
+                if (string.IsNullOrWhiteSpace(s3Folder))
+                    throw new FrameworkException("Missing S3 folder name.");
 
-                        //TODO: This locking issue might been solved on .NET 5. Removed this code after testing shows there is no issue here.
-                        //Action writeResponse = () => await context.Response.Body.Write(buffer, 0, bytesRead);
-                        //if (_detectResponseBlockingErrors)
-                        //{
-                        //    // HACK: `Response.OutputStream.Write` sometimes blocks the process at System.Web.dll!System.Web.Hosting.IIS7WorkerRequest.ExplicitFlush();
-                        //    // Until the issue is solved, this hack allows 1. logging of the problem, and 2. closing the SQL transaction while the thread remains blocked.
-                        //    // Tried removing PreSendRequestHeaders, setting aspnet:UseTaskFriendlySynchronizationContext and disabling anitivirus, but it did not help.
-                        //    // Total performance overhead of Task.Run...Wait is around 0.01 sec when downloading 200MB file with BUFFER_SIZE 100kB.
-                        //    var task = Task.Run(writeResponse);
-                        //    if (!task.Wait(_detectResponseBlockingErrorsTimeoutMs))
-                        //    {
-                        //        throw new FrameworkException(ResponseBlockedMessage +
-                        //            $" Process {Process.GetCurrentProcess().Id}, thread {Thread.CurrentThread.ManagedThreadId}," +
-                        //            $" streamed {totalBytesWritten} bytes of {file.Size}, current batch {bytesRead} bytes.");
-                        //    }
-                        //}
-                        //else
-                        //    writeResponse();
+                getObjRequest.Key = s3Folder + "/doc-" + fileMetadata.FileContentId.ToString();
 
-                        await context.Response.Body.WriteAsync(buffer, 0, bytesRead);
-                        totalBytesWritten += bytesRead;
-                        await context.Response.Body.FlushAsync();
+                try
+                {
+                    using (GetObjectResponse getObjResponse = client.GetObject(getObjRequest))
+                    {
+                        fileMetadata.Size = getObjResponse.ContentLength;
+                        Download(fileMetadata, context, getObjResponse.ResponseStream);
                     }
                 }
+                catch (Exception ex)
+                {
+                    _logger.Error("S3 storage error. Error: " + ex.ToString());
+                }
             }
-            finally
-            {
-                try { sqlTransaction.Rollback(); } catch { }
-            }
+        }
 
-            return true;
+        private void DownloadFromFileStream(FileMetadata fileMetadata, SqlConnection sqlConnection, HttpContext context)
+        {
+            using (SqlTransaction sqlTransaction = sqlConnection.BeginTransaction(IsolationLevel.ReadCommitted)) // Explicit transaction is required when working with SqlFileStream class.
+            {
+                using (var stream = SqlFileStreamProvider.GetSqlFileStreamForDownload(fileMetadata.FileContentId, sqlTransaction))
+                {
+                    Download(fileMetadata, context, stream);
+                }
+            }
         }
 
         public static readonly string ResponseBlockedMessage = $"Response.OutputStream.Write blocked.";
 
-        private async Task DownloadFromVarbinary(HttpContext context, FileMetadata file, SqlConnection sqlConnection)
+        private void DownloadFromVarbinary(FileMetadata fileMetadata, SqlConnection sqlConnection, HttpContext context)
         {
-            byte[] buffer = new byte[BUFFER_SIZE];
-
-            PopulateHeader(context, file.FileName, file.Size);
-
-            SqlCommand readCommand = new SqlCommand("SELECT Content FROM LightDMS.FileContent WHERE ID='" + file.FileContentId.ToString() + "'", sqlConnection);
-            var reader = readCommand.ExecuteReader(CommandBehavior.SequentialAccess);
-
-            while (reader.Read())
+            using (SqlCommand readCommand = new SqlCommand("SELECT Content FROM LightDMS.FileContent WHERE ID='" + fileMetadata.FileContentId.ToString() + "'", sqlConnection))
             {
-                // Read bytes into outByte[] and retain the number of bytes returned.  
-                var readed = reader.GetBytes(0, 0, buffer, 0, BUFFER_SIZE);
-                var startIndex = 0;
-                // Continue while there are bytes beyond the size of the buffer.  
-                while (readed == BUFFER_SIZE)
+                using (var sqlDataReader = readCommand.ExecuteReader(CommandBehavior.SequentialAccess))
                 {
-                    await context.Response.Body.WriteAsync(buffer, 0, (int)readed);
-                    await context.Response.Body.FlushAsync();
-
-                    // Reposition start index to end of last buffer and fill buffer.  
-                    startIndex += BUFFER_SIZE;
-                    readed = reader.GetBytes(0, startIndex, buffer, 0, BUFFER_SIZE);
+                    var success = sqlDataReader.Read();
+                    if (!success)
+                        return;
+                    Download(fileMetadata, context, sqlDataReader.GetStream(0));
                 }
-
-                await context.Response.Body.WriteAsync(buffer, 0, (int)readed);
-                await context.Response.Body.FlushAsync();
             }
-
-            reader.Close();
-            reader = null;
         }
 
         private void PopulateHeader(HttpContext context, string fileName, long length)
         {
-            _contentTypeProvider.TryGetContentType(fileName, out string contentType);
-            context.Response.ContentType = contentType;
+            context.Response.ContentType = MimeMapping.GetMimeMapping(fileName);
             // Koristiti HttpUtility.UrlPathEncode umjesto HttpUtility.UrlEncode ili Uri.EscapeDataString jer drugačije handlea SPACE i specijalne znakove
-            context.Response.Headers.Add("Content-Disposition", "attachment; filename*=UTF-8''" + HttpUtility.UrlPathEncode(fileName) + "");
-            context.Response.Headers.Add("Content-Length", length.ToString());
+            context.Response.AddHeader("Content-Disposition", "attachment; filename*=UTF-8''" + HttpUtility.UrlPathEncode(fileName) + "");
+            context.Response.StatusCode = (int)HttpStatusCode.OK;
+            context.Response.BufferOutput = false;
         }
 
         public static Guid? GetId(HttpContext context)
         {
-            var idString = context.Request.Query["id"].FirstOrDefault() ?? context.Request.Path.ToUriComponent().Split('/').Last();
+            var idString = context.Request.QueryString["id"] ?? context.Request.Url.LocalPath.Split('/').Last();
             if (!string.IsNullOrEmpty(idString) && Guid.TryParse(idString, out Guid id))
                 return id;
             else